import 2014 east coast bicycle trip

+++ drupal_extraction/BlogConverter.js
@@ -17,7 +17,7 @@
 // SELECT * FROM `node` WHERE `type` = "photo_gallery" ORDER BY `nid` DESC;
 
 const fetch = require('node-fetch');
-const {writeFileSync, mkdirSync, createWriteStream} = require('fs');
+const {existsSync, writeFileSync, mkdirSync, createWriteStream} = require('fs');
 const http = require('https');
 
 const baseUrl = "https://joshyorndorff.com";
@@ -33,8 +33,27 @@ uuids = [
 	// "540cb229-74af-41fb-bcfe-b22948a2b805", // Motorcycle
 	// "d273d557-759d-4f37-947f-aff855f7502a", // Logging
 	// "a61cd175-1221-458f-8281-9931af971cf3", // Life in Alaska
-	"18db1322-55b9-41fe-a459-feff2024b4f8", // Mini sharpie
+	// "18db1322-55b9-41fe-a459-feff2024b4f8", // Mini sharpie
 	// "2e0c32f6-e428-4c7b-b0fd-1989ba8bd6ac", // Resolutions
+
+	// 2014
+	// "6329b052-0970-465c-8b84-997830680caf", // Wild Roadtrip - Lube n Goinc
+
+	// East coast bike trip
+	"dd0a2559-73a7-4911-a48d-7acab3fd8914", // Thanks
+	"48279fd0-c266-48b9-88aa-9bc32080fd2c", // day 7
+	"739bd391-ff24-464f-9182-8b276bca1789",
+	// "0d3e752d-e610-4c97-850c-48893b756cfd", // day 5 was goofed bad. Try it again after others succeed
+	"0001b5a2-46ca-4597-8199-49dad6f66446",
+	"d1acf913-8546-44dd-bb41-c1d30526a540",
+	"bfdee918-6e89-41cf-b8d7-52ef64999e09", 
+	"5686d3ca-5a45-4d9a-97f8-b95dc29d5e81", // day 1
+	"91ae7974-857b-40a5-8e06-9f6a41b865fb", // day 0 - prep
+
+	// These two round out page two on php my admin
+	// "d77ae626-0eda-4067-9559-e7e4d1448733", // resolutions 2014
+	// "00513030-3b63-435b-8c79-0c25192c4c7d", // Flight school in Needles
+
 ];
 
 download_all(uuids);
@@ -101,7 +120,12 @@ async function downloadBlog(uuid) {
 	// Pictures will be downloaded later
 	// https://www.geeksforgeeks.org/node-js/how-to-create-a-directory-using-node-js/
 	const dashedTitle = title.replace(/\W+/g, '-');
-	mkdirSync(dashedTitle);
+	if (!existsSync(dashedTitle)){
+		mkdirSync(dashedTitle);
+	}
+	else {
+		console.log(`Directory ${dashedTitle} already exists. Not creating it`);
+	}
 
 	let contents = `+++
 title = "${title}"
@@ -116,19 +140,24 @@ ${body}
 ${photos.length > 0 ? "Photos:\n" : ""}
 `
 	for (photo of photos) {
-		console.log(`  Photo Download Started: ${photo.downloadUrl}`);
-		// Download the file from the drupal site
-		// https://stackoverflow.com/a/11944984/4184410
-		const file = createWriteStream(`${dashedTitle}/${photo.filename}`);
-		const request = http.get(photo.downloadUrl, function(response) {
-			response.pipe(file);
+		if (!existsSync(`${dashedTitle}/${photo.filename}`)){
+			console.log(`  Photo Download Started: ${photo.downloadUrl}`);
+			// Download the file from the drupal site
+			// https://stackoverflow.com/a/11944984/4184410
+			const file = createWriteStream(`${dashedTitle}/${photo.filename}`);
+			const request = http.get(photo.downloadUrl, function(response) {
+				response.pipe(file);
 
-			// after download completed close filestream
-			file.on("finish", () => {
-				file.close();
-				console.log(`  Photo Download Completed: ${photo.downloadUrl}`);
+				// after download completed close filestream
+				file.on("finish", () => {
+					file.close();
+					console.log(`  Photo Download Completed`);
+				});
 			});
-		});
+		}
+		else {
+			console.log(`Photo ${photo.downloadUrl} already exists; skipping it.`);
+		}
 
 		// Photo captions may be stored in the alt text or the title text.
 		// Possibly even slightly different versions for each :scream:
@@ -147,5 +176,10 @@ ${photos.length > 0 ? "Photos:\n" : ""}
 		}
 	}
 	
-	writeFileSync(`${dashedTitle}/index.md`, contents, {flag: "w"});	
+	if (!existsSync(`${dashedTitle}/index.md`)) {
+		writeFileSync(`${dashedTitle}/index.md`, contents, {flag: "w"});
+	}
+	else {
+		console.log("index file already exists. skipping it");
+	}
 }
